# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SkoutWatch App OpenApi API Documentation",
        "description": "SkoutWatch App Using L5 Swagger OpenApi description",
        "contact": {
            "email": "schneidershades@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://skoutwatchapi.test",
            "description": "Local API server"
        },
        {
            "url": "http://skoutwatch-dev-api.eu-west-3.elasticbeanstalk.com",
            "description": "Local API server"
        },
        {
            "url": "https://dev-api.skoutwatch.com",
            "description": "Staging API server"
        },
        {
            "url": "https://staging.skoutwatch.com",
            "description": "Staging API server"
        },
        {
            "url": "https://api.skoutwatch.com",
            "description": "Live API server"
        }
    ],
    "paths": {
...
        "/api/v1/attributes": {
            "get": {
                "tags": [
                    "Attributes"
                ],
                "summary": "Attributes of a user",
                "description": "Attributes of a user",
                "operationId": "attributes",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/api/v1/user/password/email": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "reset a registered user password",
                "description": "Returns a registered user reset email",
                "operationId": "sendResetMailForgotPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/user/password/reset": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "reset a registered user password",
                "description": "Returns a registered user reset email",
                "operationId": "resetPasswordFromToken",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/api/v1/players?per_page={per_page}": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "AllPlayer",
                "description": "Showing all player list",
                "operationId": "AllPlayer",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "path",
                        "description": "per_page",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "5"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/players": {
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "createPlayer",
                "description": "create a player and attributes",
                "operationId": "createPlayer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePlayerAttributeFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signup",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/player/{id}": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Show Player",
                "description": "Show Player",
                "operationId": "showPlayer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ForgotPasswordFormRequest": {
                "title": "Sign in Form Request Fields",
                "description": "sign in request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "title": "User reset email",
                        "description": "Reset email of the user",
                        "type": "string",
                        "example": "info@skoutwatch.com"
                    }
                },
                "type": "object"
            },
...
            "ResetPasswordRequest": {
                "title": "Sign in Form Request Fields",
                "description": "sign in request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "title": "User reset email",
                        "description": "Reset email of the user",
                        "type": "string",
                        "example": "info@skoutwatch.com"
                    },
                    "token": {
                        "title": "User token email",
                        "description": "token email of the user",
                        "type": "string",
                        "example": "nciwo3939320393dweiodwe"
                    },
                    "password": {
                        "title": "new user password",
                        "description": "new password of the user",
                        "type": "string",
                        "example": "password"
                    },
                    "password_confirmation": {
                        "title": "retype user password",
                        "description": "nretype new password of the user",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
...
            "StorePlayerAttributeFormRequest": {
                "title": "Player Create Form Request Fields",
                "description": "Player Create request body data",
                "required": [
                    "first_name"
                ],
                "properties": {
                    "first_name": {
                        "title": "first_name",
                        "description": "first_name",
                        "type": "string",
                        "example": "Schneider"
                    },
                    "last_name": {
                        "title": "last_name",
                        "description": "last_name",
                        "type": "string",
                        "example": "Komolafe"
                    },
                    "nin": {
                        "title": "nin",
                        "description": "nin",
                        "type": "string",
                        "example": "13904949033"
                    },
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string",
                        "example": "schneidershades@gmail.com"
                    },
                    "attributes": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "attribute": {
                                    "type": "string",
                                    "example": "Dribbling"
                                },
                                "score": {
                                    "type": "integer",
                                    "example": "50"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}